+++ Src/FM79979Engine/WavParser/MIDIParser/MidiFile.cpp
@@ -2107,7 +2107,7 @@ int MidiFile::getAbsoluteTickTime(double starttime) {
       buildTimeMap();
       if (timemapvalid == 0) {
          if (timemapvalid == 0) {
-            return -1.0;    // something went wrong
+            return -1;    // something went wrong
          }
       }
    }
@@ -2175,7 +2175,7 @@ int MidiFile::getTotalTimeInTicks(void) {
    if (oldTimeState == TIME_STATE_DELTA) {
       deltaTicks();
    }
-   int output = 0.0;
+   int output = 0;
    for (int i=0; i<(int)events.size(); i++) {
       if (events[i]->last().tick > output) {
          output = events[i]->last().tick;
@@ -2271,7 +2271,7 @@ int MidiFile::linearTickInterpolationAtSecond(double seconds) {
    if (timemapvalid == 0) {
       buildTimeMap();
       if (timemapvalid == 0) {
-         return -1.0;    // something went wrong
+         return -1;    // something went wrong
       }
    }
 
@@ -2324,7 +2324,7 @@ int MidiFile::linearTickInterpolationAtSecond(double seconds) {
    double y2 = timemap[startindex+1].tick;
    double xi = seconds;
 
-   return (xi-x1) * ((y2-y1)/(x2-x1)) + y1;
+   return (int)((xi-x1) * ((y2-y1)/(x2-x1)) + y1);
 }
